% Auto-generated: do not edit by hand
\name{googlePicker}

\alias{googlePicker}

\title{GooglePicker component}

\description{

}

\usage{
googlePicker(id=NULL, action=NULL, client_id=NULL, developer_key=NULL,
disabled_features=NULL, documents=NULL,
enabled_features=NULL, label=NULL, locale=NULL, open=NULL,
scope=NULL, selected_data=NULL, value=NULL, view_ids=NULL)
}

\arguments{
\item{id}{Character. }

\item{action}{Character. }

\item{client_id}{Character. }

\item{developer_key}{Character. }

\item{disabled_features}{Character | list of characters. }

\item{documents}{Named list. }

\item{enabled_features}{Character | list of characters. }

\item{label}{Character. }

\item{locale}{Character. }

\item{open}{Logical. }

\item{scope}{Character. }

\item{selected_data}{Named list. }

\item{value}{Character. }

\item{view_ids}{Character | list of characters. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

